Begin working on InstantOp running as async thread
Previously, InstantOp was running on the main thread and would deadlock the server when it ran. Ouch. Nobody likes that. Even for humor.

So far, I've begun to try implementing a multi-threaded solution based on some things I've picked up this semester. I need to look into BukkitRunnable some more, but that seems like the thing to use for this task.

//TODO Implement a mute behavior myself, why depend on the server to have a /mute command?